+++ src/SMServer/html/display.js
@@ -26,3 +26,8 @@ class Display {
 		return document.createElement(a)
 	}
 }
+
+class Photo extends Display {
+	constructor(json) {
+	}
+}