+++ docs/Changelog.md
@@ -7,7 +7,8 @@
  - Updated existing REST & Websocket API to be much more robust
  - Main view now remains responsive and shows loading indicator while starting servers/connecting to remote server
  - Fixed weird coloring in settings view ([#131](https://github.com/iandwelker/smserver/issues/131))
- - Added support in API for retrieving members of each group conversation ([#130](https://github.com/iandwelker/smserver/issues/130))
+ - Added API support for retrieving members of each group conversation ([#130](https://github.com/iandwelker/smserver/issues/130))
+ - Added API support for retrieving details about specific conversation, as opposed to just most recent conversations
 
 0.7.4 &rightarrow; 0.8.0
  - Built libsmserver into the `.deb` package so that it is no longer necessary as a dependency